set_memory_limit

+++ inception.h
@@ -9,8 +9,8 @@ private:
         string cmd_line,inf,outf;
         int uid, gid;
         string chroot_dir, working_dir, cgroup_dir;
-        int time_limit; //in ms
-        int memory_limit; //in MB
+        long long time_limit; //in ms
+        long long memory_limit; //in MB
     } architecture;
     int pid;
     char* cmd[20];
